Add colors.py + move lookup to it from scan_helpers (lookup is not scanning)
