+++ nestris_ocr/scan_strat/scan_helpers.py
@@ -1,7 +1,5 @@
 ﻿import PIL
-import numpy as np
 from nestris_ocr.config import config
-from nestris_ocr.utils.lib import ilerp
 from nestris_ocr.ocr_algo.digit import scoreImage as processDigits
 from nestris_ocr.ocr_algo.board import parseImage as processBoard
 from nestris_ocr.ocr_algo.preview2 import parseImage as processPreview
@@ -17,24 +15,6 @@ PATTERNS = {
     "stats": "DDD",
     "das": "BD",
 }
-
-REFERENCE_LEVEL_COLORS = (
-    ((0x4A, 0x32, 0xFF), (0x4A, 0xAF, 0xFE)),
-    ((0x00, 0x96, 0x00), (0x6A, 0xDC, 0x00)),
-    ((0xB0, 0x00, 0xD4), (0xFF, 0x56, 0xFF)),
-    ((0x4A, 0x32, 0xFF), (0x00, 0xE9, 0x00)),
-    ((0xC8, 0x00, 0x7F), (0x00, 0xE6, 0x78)),
-    ((0x00, 0xE6, 0x78), (0x96, 0x8D, 0xFF)),
-    ((0xC4, 0x1E, 0x0E), (0x66, 0x66, 0x66)),
-    ((0x82, 0x00, 0xFF), (0x78, 0x00, 0x41)),
-    ((0x4A, 0x32, 0xFF), (0xC4, 0x1E, 0x0E)),
-    ((0xC4, 0x1E, 0x0E), (0xF6, 0x9B, 0x00)),
-)
-
-REFERENCE_LEVEL_COLORS = [
-    (np.array(color1, dtype=np.uint8), np.array(color2, dtype=np.uint8))
-    for color1, color2 in REFERENCE_LEVEL_COLORS
-]
 
 
 # A few notes
@@ -111,25 +91,6 @@ def scan_colors(full_image):
     color2 = get_sub_image(full_image, WINDOW_AREAS["color2"])
     color2 = color2.resize((1, 1), PIL.Image.ANTIALIAS)
     color2 = color2.getpixel((0, 0))
-
-    return color1, color2
-
-
-def lookup_colors(level, black=None, white=None):  # caller must pass a valid int level
-    color1, color2 = REFERENCE_LEVEL_COLORS[level % 10]
-
-    if black and white:  # must interpolate
-        color1 = (
-            ilerp(black[0], white[0], color1[0] / 0xFF),
-            ilerp(black[1], white[1], color1[1] / 0xFF),
-            ilerp(black[2], white[2], color1[2] / 0xFF),
-        )
-
-        color2 = (
-            ilerp(black[0], white[0], color2[0] / 0xFF),
-            ilerp(black[1], white[1], color2[1] / 0xFF),
-            ilerp(black[2], white[2], color2[2] / 0xFF),
-        )
 
     return color1, color2
